+++ tw_iva_report/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [17.0.1.0.2] - 2025-12-03
+
+### Fixed
+- Added migration script to recompute VAT amounts for existing payments
+- Added `move_id.line_ids.reconciled` to compute dependencies to trigger recalculation
+
 ## [17.0.1.0.1] - 2025-12-03
 
 ### Fixed